split the web api into one router per subject
