+++ src/web/server.py
@@ -1,8 +1,8 @@
 import uvicorn
 
-import src.web.app as app_module
 from src.utils.logger import get_logger
 from src.web.app import app
+from src.web.deps import set_brain
 
 logger = get_logger("bea.web.server")
 
@@ -14,7 +14,7 @@ async def run_server(brain, host: str = "127.0.0.1", port: int = 8000):
     port on the LAN hands over full control of Bea. `--host 0.0.0.0` is a
     deliberate opt-in.
     """
-    app_module.brain_instance = brain
+    set_brain(brain)
     if host not in ("127.0.0.1", "localhost", "::1"):
         logger.warning(f"Binding on {host}: the brain API has no authentication.")
     config = uvicorn.Config(app, host=host, port=port, log_level="info")
