+++ tests/test_plan.py
@@ -363,6 +363,7 @@ def client():
     from fastapi.testclient import TestClient
 
     from src.web import app as web
+    from src.web import deps
 
     class BrainStub:
         def __init__(self):
@@ -377,12 +378,12 @@ def client():
             self.invalidations += 1
 
     stub = BrainStub()
-    previous = web.brain_instance
-    web.brain_instance = stub
+    previous = deps.brain_instance
+    deps.brain_instance = stub
     try:
         yield TestClient(web.app), stub
     finally:
-        web.brain_instance = previous
+        deps.brain_instance = previous
 
 
 def test_the_plan_starts_empty(client):
